// web/src/app/shared/api-client/model/exercise-session-dto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ExerciseDto } from './exercise-dto';
import { ExerciseSessionProgressDto } from './exercise-session-progress-dto';


export interface ExerciseSessionDto { 
    id?: string;
    type: string;
    completed?: boolean;
    flashcardDeckId?: string;
    exercise?: ExerciseDto;
    progress?: ExerciseSessionProgressDto;
}