// web/src/app/shared/api-client/model/flashcard-deck-dto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FlashcardDeckDto { 
    id?: string;
    name?: string;
    description?: string;
}