// web/src/app/shared/api-client/model/flashcard-language-dto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FlashcardLanguageDto { 
    id: string;
    abbreviation: string;
    name: string;
}